NF: all DeleteNoteMulti code together
