+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -474,37 +474,6 @@ public class CollectionTask extends BaseAsyncTask<TaskData, TaskData, TaskData> 
         }
     }
 
-
-    private static class UndoDeleteNoteMulti extends Undoable {
-        private final Note[] notesArr;
-        private final List<Card> allCards;
-
-
-        public UndoDeleteNoteMulti(Note[] notesArr, List<Card> allCards) {
-            super(Collection.DismissType.DELETE_NOTE_MULTI);
-            this.notesArr = notesArr;
-            this.allCards = allCards;
-        }
-
-
-        public long undo(Collection col) {
-            Timber.i("Undo: Delete notes");
-            // undo all of these at once instead of one-by-one
-            ArrayList<Long> ids = new ArrayList<>();
-            for (Note n : notesArr) {
-                n.flush(n.getMod(), false);
-                ids.add(n.getId());
-            }
-            for (Card c : allCards) {
-                c.flush(false);
-                ids.add(c.getId());
-            }
-            col.getDb().execute("DELETE FROM graves WHERE oid IN " + Utils.ids2str(Utils.collection2Array(ids)));
-            return MULTI_CARD;  // don't fetch new card
-
-        }
-    }
-
     
     private static class UndoChangeDeckMulti extends Undoable {
         private final Card[] cards;
@@ -648,38 +617,6 @@ public class CollectionTask extends BaseAsyncTask<TaskData, TaskData, TaskData> 
             for (Card c : cards) {
                 c.load();
             }
-            return null;
-        }
-    }
-
-    public static class DeleteNoteMulti extends DismissMulti {
-        public DeleteNoteMulti(long[] cardIds) {
-            super(cardIds);
-        }
-
-
-        public TaskData actualBackground(CollectionTask task, Card[] cards) {
-            Collection col = task.getCol();
-            AbstractSched sched = col.getSched();
-            // list of all ids to pass to remNotes method.
-            // Need Set (-> unique) so we don't pass duplicates to col.remNotes()
-            Set<Note> notes = CardUtils.getNotes(Arrays.asList(cards));
-            List<Card> allCards = CardUtils.getAllCards(notes);
-            // delete note
-            long[] uniqueNoteIds = new long[notes.size()];
-            Note[] notesArr = notes.toArray(new Note[notes.size()]);
-            int count = 0;
-            for (Note note : notes) {
-                uniqueNoteIds[count] = note.getId();
-                count++;
-            }
-
-            col.markUndo(new UndoDeleteNoteMulti(notesArr, allCards));
-
-            col.remNotes(uniqueNoteIds);
-            sched.deferReset();
-            // pass back all cards because they can't be retrieved anymore by the caller (since the note is deleted)
-            task.doProgress(new TaskData(allCards.toArray(new Card[allCards.size()])));
             return null;
         }
     }
